refactor: modularize frontend
