+++ frontend/src/App.jsx
@@ -25,6 +25,11 @@ import {
   SignUp 
 } from '@clerk/clerk-react';
 
+import { AppProvider, NotificationProvider, useApp, useNotifications } from "./contexts";
+import { Modal, ConfirmationDialog, NumberInput, ProtectedRoute, DocumentManager } from "./components";
+import { useApi, useDebounce } from "./hooks";
+import { Onboarding } from "./pages";
+
 // ===========================
 // CONFIGURATION
 // ===========================
@@ -50,372 +55,9 @@ const STRIPE_PUBLISHABLE_KEY = import.meta.env.VITE_STRIPE_PUBLISHABLE_KEY ||
 const stripePromise = loadStripe(STRIPE_PUBLISHABLE_KEY);
 
 // ===========================
-// API CLIENT
-// ===========================
-
-const createApiClient = (getToken) => {
-  const request = async (endpoint, options = {}) => {
-    try {
-      const token = await getToken();
-      
-      const config = {
-        headers: {
-          'Content-Type': 'application/json',
-          ...(token && { Authorization: `Bearer ${token}` }),
-          ...options.headers,
-        },
-        ...options,
-      };
-
-      // Don't set Content-Type for FormData
-      if (options.body instanceof FormData) {
-        delete config.headers['Content-Type'];
-      }
-
-      const response = await fetch(`${API_BASE_URL}${endpoint}`, config);
-      
-      if (!response.ok) {
-        const errorData = await response.json().catch(() => ({ error: 'Network error' }));
-        throw new Error(errorData.error || `HTTP ${response.status}`);
-      }
-
-      return await response.json();
-    } catch (error) {
-      console.error('API Error:', error);
-      throw error;
-    }
-  };
-
-  return {
-    request,
-
-    // Auth endpoints
-    getCurrentUser: () => request('/auth/me'),
-    setUserRole: (role) => request('/auth/set-role', {
-      method: 'POST',
-      body: JSON.stringify({ role }),
-    }),
-    completeProfile: (profileData) => request('/auth/complete-profile', {
-      method: 'POST',
-      body: JSON.stringify(profileData),
-    }),
-
-    // User profile endpoints
-    getUserProfile: (userId) => request(`/users/${userId}/profile`),
-    updateUserProfile: (userId, profileData) => request(`/users/${userId}/profile`, {
-      method: 'PUT',
-      body: JSON.stringify(profileData),
-    }),
-
-    // Project endpoints
-    getProjects: () => request('/projects'),
-    createProject: (projectData) => request('/projects', {
-      method: 'POST',
-      body: JSON.stringify(projectData),
-    }),
-    getProject: (id) => request(`/projects/${id}`),
-    updateProject: (id, projectData) => request(`/projects/${id}`, {
-      method: 'PUT',
-      body: JSON.stringify(projectData),
-    }),
-
-    // Document endpoints
-    uploadDocuments: (projectId, formData) => request(`/projects/${projectId}/documents`, {
-      method: 'POST',
-      body: formData,
-    }),
-    getProjectDocuments: (projectId) => request(`/projects/${projectId}/documents`),
-    deleteDocument: (documentId) => request(`/documents/${documentId}`, {
-      method: 'DELETE',
-    }),
-    getRequiredDocuments: () => request('/required-documents'),
-    
-    // Document download
-    async downloadDocument(filePath) {
-  const token = await getToken();
-  // Extract just the filename from any path format
-  const filename = filePath.split('/').pop();
-  const response = await fetch(`${API_BASE_URL.replace('/api', '')}/uploads/${filename}`, {
-    headers: {
-      Authorization: `Bearer ${token}`
-    }
-  });
-  
-  if (!response.ok) throw new Error('Download failed');
-  return response.blob();
-},
-
-// Add these to your api client object
-getDeal: (dealId) => request(`/deals/${dealId}`),
-createDeal: (projectId, accessRequestId) => request('/deals', {
-  method: 'POST',
-  body: JSON.stringify({ project_id: projectId, access_request_id: accessRequestId }),
-}),
-getDealDocuments: (dealId) => request(`/deals/${dealId}/documents`),
-uploadDealDocuments: (dealId, formData) => request(`/deals/${dealId}/documents`, {
-  method: 'POST',
-  body: formData,
-}),
-getDocumentRequests: (dealId) => request(`/deals/${dealId}/document-requests`),
-createDocumentRequest: (dealId, requestData) => request(`/deals/${dealId}/document-requests`, {
-  method: 'POST',
-  body: JSON.stringify(requestData),
-}),
-downloadDealDocument: (dealId, documentId) => request(`/deals/${dealId}/documents/${documentId}`),
-
-    // Payment endpoints
-    createProjectPayment: (projectId) => request('/payments/create-project-payment', {
-      method: 'POST',
-      body: JSON.stringify({ project_id: projectId }),
-    }),
-    createSubscription: (paymentMethodId) => request('/payments/create-subscription', {
-      method: 'POST',
-      body: JSON.stringify({ payment_method_id: paymentMethodId }),
-    }),
-    cancelSubscription: () => request('/payments/cancel-subscription', {
-      method: 'POST',
-    }),
-    
-    // For development/testing
-    simulatePaymentSuccess: (projectId, paymentIntentId) => request('/payments/simulate-success', {
-      method: 'POST',
-      body: JSON.stringify({ project_id: projectId, payment_intent_id: paymentIntentId }),
-    }),
-    simulateSubscription: () => request('/payments/simulate-subscription', {
-      method: 'POST',
-    }),
-
-    // Access request endpoints
-    requestAccess: (projectId, initialMessage) => request('/access-requests', {
-      method: 'POST',
-      body: JSON.stringify({ project_id: projectId, initial_message: initialMessage }),
-    }),
-    getAccessRequests: () => request('/access-requests'),
-    approveAccessRequest: (requestId) => request(`/access-requests/${requestId}/approve`, {
-      method: 'PUT',
-    }),
-    declineAccessRequest: (requestId) => request(`/access-requests/${requestId}/decline`, {
-      method: 'PUT',
-    }),
-
-    // Messaging endpoints
-    getMessages: (requestId) => request(`/access-requests/${requestId}/messages`),
-    sendMessage: (requestId, message) => request(`/access-requests/${requestId}/messages`, {
-      method: 'POST',
-      body: JSON.stringify({ message, message_type: 'text' }),
-    }),
-    markMessageAsRead: (messageId) => request(`/messages/${messageId}/read`, {
-      method: 'PUT',
-    }),
-
-    // AI Chat endpoints
-    createAIChatSession: (projectId, sessionTitle) => request('/ai-chat/sessions', {
-      method: 'POST',
-      body: JSON.stringify({ project_id: projectId, session_title: sessionTitle }),
-    }),
-    getAIChatSessions: () => request('/ai-chat/sessions'),
-    getAIChatMessages: (sessionId) => request(`/ai-chat/sessions/${sessionId}/messages`),
-    sendAIChatMessage: (sessionId, message) => request(`/ai-chat/sessions/${sessionId}/messages`, {
-      method: 'POST',
-      body: JSON.stringify({ message }),
-    }),
-
-    // Admin endpoints
-    getUsers: () => request('/admin/users'),
-    approveUser: (userId) => request(`/admin/users/${userId}/approve`, {
-      method: 'PUT',
-    }),
-    getAdminStats: () => request('/admin/stats'),
-    getSystemSettings: () => request('/admin/system-settings'),
-    updateSystemSetting: (key, value) => request(`/admin/system-settings/${key}`, {
-      method: 'PUT',
-      body: JSON.stringify({ value }),
-    }),
-
-    // Notification preferences
-    getNotificationPreferences: () => request('/notifications/preferences'),
-    updateNotificationPreferences: (preferences) => request('/notifications/preferences', {
-      method: 'PUT',
-      body: JSON.stringify(preferences),
-    }),
-
-    // Data export
-    exportUserData: () => request('/export/user-data'),
-    exportProjects: () => request('/export/projects'),
-    
-    // Account management
-    deleteAccount: (confirmation) => request('/account/delete', {
-      method: 'DELETE',
-      body: JSON.stringify({ confirmation }),
-    }),
-  };
-};
 
 // ===========================
 // CONTEXTS
-// ===========================
-
-// App Context
-const AppContext = createContext();
-
-const useApp = () => {
-  const context = useContext(AppContext);
-  if (!context) {
-    throw new Error('useApp must be used within AppProvider');
-  }
-  return context;
-};
-
-// Notification Context
-const NotificationContext = createContext();
-
-const useNotifications = () => {
-  const context = useContext(NotificationContext);
-  if (!context) {
-    throw new Error('useNotifications must be used within NotificationProvider');
-  }
-  return context;
-};
-
-// ===========================
-// PROVIDERS
-// ===========================
-
-const AppProvider = ({ children }) => {
-  const { user: clerkUser, isSignedIn, isLoaded } = useUser();
-  const { getToken } = useAuth();
-  const [userData, setUserData] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
-  const fetchUserData = useCallback(async () => {
-    if (!isSignedIn || !clerkUser) {
-      setUserData(null);
-      setLoading(false);
-      return;
-    }
-
-    try {
-      const api = createApiClient(getToken);
-      const data = await api.getCurrentUser();
-      setUserData(data.user);
-      setError(null);
-    } catch (err) {
-      console.error('Failed to fetch user data:', err);
-      setError(err.message);
-      // Don't clear userData on error - keep cached version
-    } finally {
-      setLoading(false);
-    }
-  }, [isSignedIn, clerkUser, getToken]);
-
-  useEffect(() => {
-    if (isLoaded) {
-      fetchUserData();
-    }
-  }, [isLoaded, fetchUserData]);
-
-  const refreshUser = async () => {
-    await fetchUserData();
-  };
-
-  const value = {
-    user: userData,
-    loading,
-    error,
-    refreshUser,
-    isAuthenticated: isSignedIn
-  };
-
-  return (
-    <AppContext.Provider value={value}>
-      {children}
-    </AppContext.Provider>
-  );
-};
-
-const NotificationProvider = ({ children }) => {
-  const [notifications, setNotifications] = useState([]);
-  const [unreadCount, setUnreadCount] = useState(0);
-
-  const addNotification = (notification) => {
-    const id = Date.now();
-    const newNotification = {
-      id,
-      ...notification,
-      timestamp: new Date(),
-      read: false
-    };
-    
-    setNotifications(prev => [newNotification, ...prev]);
-    setUnreadCount(prev => prev + 1);
-    
-    // Auto-remove after 5 seconds if it's a toast
-    if (notification.type === 'toast') {
-      setTimeout(() => {
-        removeNotification(id);
-      }, 5000);
-    }
-  };
-
-  const removeNotification = (id) => {
-    setNotifications(prev => prev.filter(n => n.id !== id));
-  };
-
-  const markAsRead = (id) => {
-    setNotifications(prev => 
-      prev.map(n => n.id === id ? { ...n, read: true } : n)
-    );
-    setUnreadCount(prev => Math.max(0, prev - 1));
-  };
-
-  const markAllAsRead = () => {
-    setNotifications(prev => prev.map(n => ({ ...n, read: true })));
-    setUnreadCount(0);
-  };
-
-  const value = {
-    notifications,
-    unreadCount,
-    addNotification,
-    removeNotification,
-    markAsRead,
-    markAllAsRead
-  };
-
-  return (
-    <NotificationContext.Provider value={value}>
-      {children}
-    </NotificationContext.Provider>
-  );
-};
-
-// ===========================
-// HOOKS
-// ===========================
-
-const useApi = () => {
-  const { getToken } = useAuth();
-  return createApiClient(getToken);
-};
-
-const useDebounce = (value, delay) => {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-};
-
-// ===========================
 // UTILITY FUNCTIONS
 // ===========================
 
@@ -518,7 +160,7 @@ const validateABN = (abn) => {
 // ===========================
 
 // Loading Spinner
-const LoadingSpinner = ({ size = 'default', message }) => (
+export const LoadingSpinner = ({ size = 'default', message }) => (
   <div className={`loading-spinner ${size}`}>
     <div className="spinner"></div>
     {message && <p className="loading-message">{message}</p>}
@@ -567,7 +209,7 @@ const Toast = ({ message, type = 'info', onClose }) => {
 };
 
 // Enhanced Tooltip Component
-const Tooltip = ({ children, content, position = 'top' }) => {
+export const Tooltip = ({ children, content, position = 'top' }) => {
   return (
     <div className="tooltip-wrapper">
       {children}
@@ -579,141 +221,6 @@ const Tooltip = ({ children, content, position = 'top' }) => {
 };
 
 // Modal
-const Modal = ({ isOpen, onClose, title, children, size = 'medium' }) => {
-  useEffect(() => {
-    if (isOpen) {
-      document.body.style.overflow = 'hidden';
-    } else {
-      document.body.style.overflow = 'unset';
-    }
-    
-    return () => {
-      document.body.style.overflow = 'unset';
-    };
-  }, [isOpen]);
-
-  if (!isOpen) return null;
-
-  return (
-    <div className="modal-overlay" onClick={onClose}>
-      <div className={`modal-content modal-${size}`} onClick={e => e.stopPropagation()}>
-        <div className="modal-header">
-          <h2>{title}</h2>
-          <button onClick={onClose} className="close-btn">×</button>
-        </div>
-        <div className="modal-body">
-          {children}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-// Confirmation Dialog
-const ConfirmationDialog = ({ isOpen, onClose, onConfirm, title, message, confirmText = 'Confirm', cancelText = 'Cancel', danger = false }) => {
-  const [inputValue, setInputValue] = useState('');
-  const requiresInput = message.includes('type "DELETE"');
-  
-  const handleConfirm = () => {
-    if (requiresInput && inputValue !== 'DELETE') {
-      return;
-    }
-    onConfirm();
-    onClose();
-  };
-
-  return (
-    <Modal isOpen={isOpen} onClose={onClose} title={title} size="small">
-      <div className="confirmation-dialog">
-        <p>{message}</p>
-        {requiresInput && (
-          <input
-            type="text"
-            value={inputValue}
-            onChange={(e) => setInputValue(e.target.value)}
-            placeholder="Type DELETE to confirm"
-            className="form-input"
-          />
-        )}
-        <div className="dialog-actions">
-          <button onClick={onClose} className="btn btn-outline">
-            {cancelText}
-          </button>
-          <button 
-            onClick={handleConfirm} 
-            className={`btn ${danger ? 'btn-danger' : 'btn-primary'}`}
-            disabled={requiresInput && inputValue !== 'DELETE'}
-          >
-            {confirmText}
-          </button>
-        </div>
-      </div>
-    </Modal>
-  );
-};
-
-// Number Input with Format
-const NumberInput = ({ value, onChange, placeholder, prefix = '', suffix = '', min, max, step = 1, disabled = false }) => {
-  const inputRef = useRef(null);
-  const [displayValue, setDisplayValue] = useState('');
-  const [isFocused, setIsFocused] = useState(false);
-
-  useEffect(() => {
-    if (!isFocused && value) {
-      setDisplayValue(formatNumber(value));
-    } else if (!isFocused) {
-      setDisplayValue('');
-    }
-  }, [value, isFocused]);
-
-  const handleFocus = () => {
-    setIsFocused(true);
-    setDisplayValue(value || '');
-    // Prevent scroll
-    if (inputRef.current) {
-      inputRef.current.addEventListener('wheel', preventDefault, { passive: false });
-    }
-  };
-
-  const handleBlur = () => {
-    setIsFocused(false);
-    if (inputRef.current) {
-      inputRef.current.removeEventListener('wheel', preventDefault);
-    }
-  };
-
-  const preventDefault = (e) => {
-    e.preventDefault();
-  };
-
-  const handleChange = (e) => {
-    const rawValue = e.target.value.replace(/[^0-9.-]/g, '');
-    setDisplayValue(rawValue);
-    
-    const numValue = parseFloat(rawValue) || 0;
-    if (onChange) {
-      onChange(numValue);
-    }
-  };
-
-  return (
-    <div className="number-input-wrapper">
-      {prefix && <span className="input-prefix">{prefix}</span>}
-      <input
-        ref={inputRef}
-        type="text"
-        value={displayValue}
-        onChange={handleChange}
-        onFocus={handleFocus}
-        onBlur={handleBlur}
-        placeholder={placeholder}
-        className="form-input number-input"
-        disabled={disabled}
-      />
-      {suffix && <span className="input-suffix">{suffix}</span>}
-    </div>
-  );
-};
 
 // Progress Bar
 const ProgressBar = ({ value, max = 100, label, showPercentage = true }) => {
@@ -1175,28 +682,6 @@ const BrokerAIFloating = () => {
 // PROTECTED ROUTE
 // ===========================
 
-const ProtectedRoute = ({ children, roles = [] }) => {
-  const { user, loading, isAuthenticated } = useApp();
-  const location = useLocation();
-
-  if (loading) {
-    return <LoadingSpinner message="Loading..." />;
-  }
-
-  if (!isAuthenticated) {
-    return <Navigate to="/login" state={{ from: location }} replace />;
-  }
-
-  if (!user || !user.role) {
-    return <Navigate to="/onboarding" replace />;
-  }
-
-  if (roles.length > 0 && !roles.includes(user.role)) {
-    return <Navigate to="/dashboard" replace />;
-  }
-
-  return children;
-};
 
 // ===========================
 // AUTH PAGES
@@ -1291,403 +776,6 @@ const ClerkAuthWrapper = ({ mode }) => {
 
 // ===========================
 // ONBOARDING
-// ===========================
-
-const Onboarding = () => {
-  const { user: clerkUser } = useUser();
-  const { refreshUser } = useApp();
-  const navigate = useNavigate();
-  const api = useApi();
-  const [step, setStep] = useState('role');
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState('');
-  const [formData, setFormData] = useState({
-    role: '',
-    // Funder profile fields
-    company_name: '',
-    company_type: '',
-    investment_focus: '',
-    typical_deal_size_min: '',
-    typical_deal_size_max: '',
-    years_experience: '',
-    aum: '',
-    phone: '',
-    linkedin: '',
-    bio: '',
-    abn: ''
-  });
-
-  const handleRoleSelection = async (role) => {
-    setLoading(true);
-    setError('');
-    
-    try {
-      await api.setUserRole(role);
-      
-      if (role === 'borrower') {
-        await refreshUser();
-        navigate('/dashboard');
-      } else {
-        setFormData({ ...formData, role });
-        setStep('profile');
-      }
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const validateProfile = () => {
-    const errors = [];
-    
-    if (!formData.company_name) errors.push('Company name is required');
-    if (!formData.company_type) errors.push('Company type is required');
-    if (!formData.investment_focus) errors.push('Investment focus is required');
-    if (!formData.typical_deal_size_min) errors.push('Minimum deal size is required');
-    if (!formData.typical_deal_size_max) errors.push('Maximum deal size is required');
-    if (!formData.years_experience) errors.push('Years of experience is required');
-    if (!formData.phone) errors.push('Phone number is required');
-    if (!formData.abn) errors.push('ABN is required');
-    
-    if (!validatePhone(formData.phone)) errors.push('Invalid phone number format');
-    if (!validateABN(formData.abn)) errors.push('Invalid ABN format');
-    
-    if (parseInt(formData.typical_deal_size_min) >= parseInt(formData.typical_deal_size_max)) {
-      errors.push('Maximum deal size must be greater than minimum');
-    }
-    
-    return errors;
-  };
-
-  const handleProfileSubmit = async (e) => {
-    e.preventDefault();
-    
-    const errors = validateProfile();
-    if (errors.length > 0) {
-      setError(errors.join(', '));
-      return;
-    }
-    
-    setLoading(true);
-    setError('');
-
-    try {
-      await api.completeProfile({
-        ...formData,
-        typical_deal_size_min: parseInt(formData.typical_deal_size_min),
-        typical_deal_size_max: parseInt(formData.typical_deal_size_max),
-        years_experience: parseInt(formData.years_experience),
-        aum: formData.aum ? parseInt(formData.aum) : null
-      });
-      await refreshUser();
-      setStep('complete');
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  if (step === 'role') {
-    return (
-      <div className="onboarding-container">
-        <div className="onboarding-card">
-          <h1>Welcome to Tranch</h1>
-          <p>Let's get you set up. Are you a property developer or an investor?</p>
-          
-          {error && <ErrorMessage message={error} onClose={() => setError('')} />}
-          
-          <div className="role-selection">
-            <button 
-              className="role-card"
-              onClick={() => handleRoleSelection('borrower')}
-              disabled={loading}
-            >
-              <div className="role-icon">🏗</div>
-              <h3>I'm a Developer</h3>
-              <p>I need funding for property development projects</p>
-            </button>
-            
-            <button 
-              className="role-card"
-              onClick={() => handleRoleSelection('funder')}
-              disabled={loading}
-            >
-              <div className="role-icon">💰</div>
-              <h3>I'm an Investor</h3>
-              <p>I want to invest in property development projects</p>
-            </button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  if (step === 'profile') {
-    return (
-      <div className="onboarding-container">
-        <div className="onboarding-card wide">
-          <h1>Complete Your Investor Profile</h1>
-          <p>This information helps developers understand your investment criteria</p>
-          
-          {error && <ErrorMessage message={error} onClose={() => setError('')} />}
-          
-          <form onSubmit={handleProfileSubmit} className="onboarding-form">
-            <div className="form-section">
-              <h3>Company Information</h3>
-              
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="company_name">
-                    Company Name *
-                    <Tooltip content="Your registered business name">
-                      <span className="help-icon">?</span>
-                    </Tooltip>
-                  </label>
-                  <input
-                    type="text"
-                    id="company_name"
-                    value={formData.company_name}
-                    onChange={(e) => setFormData({ ...formData, company_name: e.target.value })}
-                    required
-                    className="form-input"
-                    placeholder="ABC Capital Partners"
-                  />
-                </div>
-
-                <div className="form-group">
-                  <label htmlFor="abn">
-                    ABN *
-                    <Tooltip content="11-digit Australian Business Number">
-                      <span className="help-icon">?</span>
-                    </Tooltip>
-                  </label>
-                  <input
-                    type="text"
-                    id="abn"
-                    value={formData.abn}
-                    onChange={(e) => setFormData({ ...formData, abn: e.target.value })}
-                    required
-                    className="form-input"
-                    placeholder="12 345 678 901"
-                    maxLength="14"
-                  />
-                </div>
-              </div>
-
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="company_type">
-                    Company Type *
-                    <Tooltip content="Select the category that best describes your organization">
-                      <span className="help-icon">?</span>
-                    </Tooltip>
-                  </label>
-                  <select
-                    id="company_type"
-                    value={formData.company_type}
-                    onChange={(e) => setFormData({ ...formData, company_type: e.target.value })}
-                    className="form-select"
-                    required
-                  >
-                    <option value="">Select company type</option>
-                    <option value="Private Credit Fund">Private Credit Fund</option>
-                    <option value="Investment Bank">Investment Bank</option>
-                    <option value="Family Office">Family Office</option>
-                    <option value="Hedge Fund">Hedge Fund</option>
-                    <option value="Real Estate Fund">Real Estate Fund</option>
-                    <option value="High Net Worth Individual">High Net Worth Individual</option>
-                    <option value="Other">Other</option>
-                  </select>
-                </div>
-
-                <div className="form-group">
-                  <label htmlFor="years_experience">
-                    Years Experience *
-                    <Tooltip content="Years of experience in property investment">
-                      <span className="help-icon">?</span>
-                    </Tooltip>
-                  </label>
-                  <NumberInput
-                    id="years_experience"
-                    value={formData.years_experience}
-                    onChange={(value) => setFormData({ ...formData, years_experience: value })}
-                    placeholder="10"
-                    min={0}
-                    max={100}
-                  />
-                </div>
-              </div>
-            </div>
-
-            <div className="form-section">
-              <h3>Investment Profile</h3>
-              
-              <div className="form-group">
-                <label htmlFor="investment_focus">
-                  Investment Focus *
-                  <Tooltip content="Primary property types you invest in">
-                    <span className="help-icon">?</span>
-                  </Tooltip>
-                </label>
-                <select
-                  id="investment_focus"
-                  value={formData.investment_focus}
-                  onChange={(e) => setFormData({ ...formData, investment_focus: e.target.value })}
-                  className="form-select"
-                  required
-                >
-                  <option value="">Select investment focus</option>
-                  <option value="Residential Development">Residential Development</option>
-                  <option value="Commercial Development">Commercial Development</option>
-                  <option value="Mixed-Use Development">Mixed-Use Development</option>
-                  <option value="Industrial Development">Industrial Development</option>
-                  <option value="All Property Types">All Property Types</option>
-                </select>
-              </div>
-
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="typical_deal_size_min">
-                    Min Deal Size (AUD) *
-                    <Tooltip content="Minimum investment amount per deal">
-                      <span className="help-icon">?</span>
-                    </Tooltip>
-                  </label>
-                  <NumberInput
-                    id="typical_deal_size_min"
-                    value={formData.typical_deal_size_min}
-                    onChange={(value) => setFormData({ ...formData, typical_deal_size_min: value })}
-                    placeholder="1,000,000"
-                    prefix="$"
-                    min={1}
-                  />
-                </div>
-
-                <div className="form-group">
-                  <label htmlFor="typical_deal_size_max">
-                    Max Deal Size (AUD) *
-                    <Tooltip content="Maximum investment amount per deal">
-                      <span className="help-icon">?</span>
-                    </Tooltip>
-                  </label>
-                  <NumberInput
-                    id="typical_deal_size_max"
-                    value={formData.typical_deal_size_max}
-                    onChange={(value) => setFormData({ ...formData, typical_deal_size_max: value })}
-                    placeholder="50,000,000"
-                    prefix="$"
-                    min={1}
-                  />
-                </div>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="aum">
-                  Assets Under Management (AUD)
-                  <Tooltip content="Total assets your organization manages (optional)">
-                    <span className="help-icon">?</span>
-                  </Tooltip>
-                </label>
-                <NumberInput
-                  id="aum"
-                  value={formData.aum}
-                  onChange={(value) => setFormData({ ...formData, aum: value })}
-                  placeholder="100,000,000"
-                  prefix="$"
-                  min={0}
-                />
-              </div>
-            </div>
-
-            <div className="form-section">
-              <h3>Contact Information</h3>
-              
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="phone">
-                    Phone Number *
-                    <Tooltip content="Australian mobile or landline">
-                      <span className="help-icon">?</span>
-                    </Tooltip>
-                  </label>
-                  <input
-                    type="tel"
-                    id="phone"
-                    value={formData.phone}
-                    onChange={(e) => setFormData({ ...formData, phone: e.target.value })}
-                    required
-                    className="form-input"
-                    placeholder="+61 400 000 000"
-                  />
-                </div>
-
-                <div className="form-group">
-                  <label htmlFor="linkedin">
-                    LinkedIn Profile
-                    <Tooltip content="Your professional LinkedIn URL (optional)">
-                      <span className="help-icon">?</span>
-                    </Tooltip>
-                  </label>
-                  <input
-                    type="url"
-                    id="linkedin"
-                    value={formData.linkedin}
-                    onChange={(e) => setFormData({ ...formData, linkedin: e.target.value })}
-                    className="form-input"
-                    placeholder="https://linkedin.com/in/yourprofile"
-                  />
-                </div>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="bio">
-                  Professional Bio
-                  <Tooltip content="Brief description of your investment philosophy and experience">
-                    <span className="help-icon">?</span>
-                  </Tooltip>
-                </label>
-                <textarea
-                  id="bio"
-                  value={formData.bio}
-                  onChange={(e) => setFormData({ ...formData, bio: e.target.value })}
-                  className="form-textarea"
-                  placeholder="Brief professional background and investment philosophy..."
-                  rows="4"
-                  maxLength="500"
-                />
-                <div className="character-count">{formData.bio.length}/500</div>
-              </div>
-            </div>
-
-            <button type="submit" disabled={loading} className="btn btn-primary btn-block">
-              {loading ? 'Submitting...' : 'Complete Profile'}
-            </button>
-          </form>
-        </div>
-      </div>
-    );
-  }
-
-  if (step === 'complete') {
-    return (
-      <div className="onboarding-container">
-        <div className="onboarding-card">
-          <div className="success-icon">✓</div>
-          <h1>Profile Submitted!</h1>
-          <p>Your profile is under review. We'll notify you once approved (usually within 24 hours).</p>
-          <button onClick={() => navigate('/dashboard')} className="btn btn-primary">
-            Go to Dashboard
-          </button>
-        </div>
-      </div>
-    );
-  }
-};
-
-// ===========================
-// DASHBOARD
 // ===========================
 
 const Dashboard = () => {
@@ -5586,7 +4674,7 @@ const DealRoom = () => {
       <div className="deal-content">
         {activeTab === 'overview' && <DealOverview deal={deal} project={project} />}
         {activeTab === 'documents' && (
-          <DealDocumentManager 
+          <DocumentManager 
             dealId={dealId}
             userRole={user.role}
             onUpdate={fetchDealData}
@@ -5681,132 +4769,6 @@ const DealOverview = ({ deal, project }) => {
             </div>
           </div>
         </div>
-      </div>
-    </div>
-  );
-};
-
-const DealDocumentManager = ({ dealId, userRole, onUpdate }) => {
-  const [documents, setDocuments] = useState([]);
-  const [requests, setRequests] = useState([]);
-  const [uploading, setUploading] = useState(false);
-  
-  useEffect(() => {
-    // Simulate document data for now
-    setDocuments([
-      {
-        id: 1,
-        file_name: 'Financial_Analysis.pdf',
-        uploader_name: 'Developer',
-        uploaded_at: new Date().toISOString(),
-        file_type: 'pdf'
-      },
-      {
-        id: 2,
-        file_name: 'Construction_Contract.pdf',
-        uploader_name: 'Developer',
-        uploaded_at: new Date().toISOString(),
-        file_type: 'pdf'
-      }
-    ]);
-    
-    setRequests([
-      {
-        id: 1,
-        document_name: 'Insurance Certificate',
-        description: 'Current insurance certificate for the project',
-        requester_name: 'Funder',
-        requester_role: 'funder',
-        created_at: new Date().toISOString()
-      }
-    ]);
-  }, [dealId]);
-  
-  const handleUpload = async (files, requestId = null) => {
-    setUploading(true);
-    
-    // Simulate upload
-    setTimeout(() => {
-      const newDoc = {
-        id: Date.now(),
-        file_name: files[0].name,
-        uploader_name: userRole === 'funder' ? 'Funder' : 'Developer',
-        uploaded_at: new Date().toISOString(),
-        file_type: 'pdf'
-      };
-      setDocuments(prev => [...prev, newDoc]);
-      setUploading(false);
-      if (onUpdate) onUpdate();
-    }, 1000);
-  };
-  
-  return (
-    <div className="document-manager">
-      <div className="document-header">
-        <h3>Deal Documents</h3>
-        <div className="document-actions">
-          <button className="btn btn-outline">
-            Request Document
-          </button>
-          <label className="btn btn-primary">
-            Upload Documents
-            <input
-              type="file"
-              multiple
-              onChange={(e) => handleUpload(Array.from(e.target.files))}
-              style={{ display: 'none' }}
-            />
-          </label>
-        </div>
-      </div>
-      
-      {requests.length > 0 && (
-        <div className="document-requests">
-          <h4>Outstanding Requests</h4>
-          {requests.map(request => (
-            <div key={request.id} className="request-card">
-              <div className="request-info">
-                <h5>{request.document_name}</h5>
-                <p>{request.description}</p>
-                <span className="request-meta">
-                  Requested by {request.requester_name} • {formatDate(request.created_at)}
-                </span>
-              </div>
-              {request.requester_role !== userRole && (
-                <label className="btn btn-sm btn-primary">
-                  Upload Response
-                  <input
-                    type="file"
-                    onChange={(e) => handleUpload(Array.from(e.target.files), request.id)}
-                    style={{ display: 'none' }}
-                  />
-                </label>
-              )}
-            </div>
-          ))}
-        </div>
-      )}
-      
-      <div className="documents-grid">
-        {documents.map(doc => (
-          <div key={doc.id} className="document-card">
-            <div className="doc-icon">
-              <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                <path d="M7 21h10a2 2 0 002-2V9.414a1 1 0 00-.293-.707l-5.414-5.414A1 1 0 0012.586 3H7a2 2 0 00-2 2v14a2 2 0 002 2z" />
-              </svg>
-            </div>
-            <div className="doc-info">
-              <h4>{doc.file_name}</h4>
-              <p>Uploaded by {doc.uploader_name}</p>
-              <span className="doc-meta">{formatDate(doc.uploaded_at)}</span>
-            </div>
-            <div className="doc-actions">
-              <button className="btn btn-sm btn-outline">
-                Download
-              </button>
-            </div>
-          </div>
-        ))}
       </div>
     </div>
   );
