+++ frontend/src/components/DocumentManager.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { useApi } from '../../hooks/useApi';
-import { formatDate } from '../../utils/formatters';
+import { useApi } from '../hooks/useApi';
+import { formatDate } from '../utils/formatters';
 
 const DocumentManager = ({ dealId, userRole, onUpdate }) => {
   const api = useApi();
@@ -138,4 +138,4 @@ const DocumentManager = ({ dealId, userRole, onUpdate }) => {
   );
 };
 
-export default DocumentManager;
+export { DocumentManager };
