+++ frontend/src/components/index.js
@@ -1,0 +1,4 @@
+export * from './DocumentManager';
+export * from './Modal';
+export * from './NumberInput';
+export * from './ProtectedRoute';
