+++ frontend/src/contexts/index.js
@@ -1,0 +1,2 @@
+export * from './AppContext';
+export * from './NotificationContext';
